refactor: removed shared folder and moved user interfaces to corresponding folder
