+++ frontend/src/app/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './authentication';
+export * from './user';
